Modify code to analyze series of frames
Modify code to take in a list of frames and analyze each of them using
the emotion API. Created a queue of tasks populated by a producer,
which creates a task to await the query results from each frame. A
consumer then goes through the queue tasks sequentially, waits on them
until completion, and prints out the analysis result.