Make video analysis faster
Create a producer and consumer for video frames. Producer extracts
frames from video, makes an async call to cognitive face api, and
places result in a queue. We don’t wait for the result however. The
consumer waits to get the next item from the queue and processes the
result. This ensures frames will be processed in order and more
efficient.